feat: modified login and sign up page and create forgot and reset password page
