+++ src/App.jsx
@@ -2,7 +2,10 @@ import { createBrowserRouter, Form } from "react-router";
 import { RouterProvider } from "react-router/dom";
 import GuestLayout from "./components/layouts/GuestLayout";
 import RootLayout from "./components/layouts/RootLayout";
-import Authentication from "./pages/Login/Authentication";
+import Login from "./pages/Authentication/Login";
+import Signup from "./pages/Authentication/Signup";
+import ForgotPassword from "./pages/Authentication/ForgotPassword";
+import ResetPassword from "./pages/Authentication/ResetPassword";
 
 const router = createBrowserRouter([
   {
@@ -14,7 +17,12 @@ const router = createBrowserRouter([
         children: [
           {
             path: "auth",
-            children: [{ path: ":page", Component: Authentication }],
+            children: [
+              { path: "login", Component: Login },
+              { path: "signup", Component: Signup },
+              { path: "forgot-password", Component: ForgotPassword },
+              { path: "reset-password", Component: ResetPassword },
+            ],
           },
         ],
       },
